+++ internal/cachecontroller/cache_controller.go
@@ -17,7 +17,6 @@ import (
 
 	"github.com/openfga/openfga/internal/build"
 	"github.com/openfga/openfga/internal/concurrency"
-	"github.com/openfga/openfga/internal/utils"
 	"github.com/openfga/openfga/pkg/logger"
 	"github.com/openfga/openfga/pkg/storage"
 	"github.com/openfga/openfga/pkg/telemetry"
@@ -57,8 +56,11 @@ var (
 )
 
 type CacheController interface {
-	// DetermineInvalidationTime returns the timestamp of the last write for the specified store if it was in cache,
-	// Else it returns Zero time and triggers InvalidateIfNeeded().
+	// DetermineInvalidationTime returns the timestamp of the last write for the
+	// specified store if it was in cache, else it returns the Zero time and
+	// triggers InvalidateIfNeeded(). The last write time can be used to determine
+	// whether a cached entry is stil valid - if it was cached before the last write
+	// to the store, it can't be trusted anymore.
 	DetermineInvalidationTime(context.Context, string) time.Time
 
 	// InvalidateIfNeeded checks to see if an invalidation is currently in progress for a store,
@@ -99,8 +101,10 @@ type InMemoryCacheController struct {
 	ds    storage.OpenFGADatastore
 	cache storage.InMemoryCache[any]
 
-	// ttl for the entry that keeps the last timestamp for a Write for a storeID.
+	// ttl is the minimum time interval to trigger cache invalidation for
+	// a given store.
 	ttl                   time.Duration
+	queryCacheTTL         time.Duration
 	iteratorCacheTTL      time.Duration
 	inflightInvalidations sync.Map
 	logger                logger.Logger
@@ -109,11 +113,19 @@ type InMemoryCacheController struct {
 	wg sync.WaitGroup
 }
 
-func NewCacheController(ds storage.OpenFGADatastore, cache storage.InMemoryCache[any], ttl time.Duration, iteratorCacheTTL time.Duration, opts ...InMemoryCacheControllerOpt) CacheController {
+func NewCacheController(
+	ds storage.OpenFGADatastore,
+	cache storage.InMemoryCache[any],
+	ttl time.Duration,
+	queryCacheTTL time.Duration,
+	iteratorCacheTTL time.Duration,
+	opts ...InMemoryCacheControllerOpt,
+) CacheController {
 	c := &InMemoryCacheController{
 		ds:                    ds,
 		cache:                 cache,
 		ttl:                   ttl,
+		queryCacheTTL:         queryCacheTTL,
 		iteratorCacheTTL:      iteratorCacheTTL,
 		inflightInvalidations: sync.Map{},
 		logger:                logger.NewNoopLogger(),
@@ -126,8 +138,11 @@ func NewCacheController(ds storage.OpenFGADatastore, cache storage.InMemoryCache
 	return c
 }
 
-// DetermineInvalidationTime returns the timestamp of the last write for the specified store if it was in cache,
-// Else it returns Zero time and triggers InvalidateIfNeeded().
+// DetermineInvalidationTime returns the timestamp of the last write for the
+// specified store if it was in cache, else it returns the Zero time and
+// triggers InvalidateIfNeeded(). The last write time can be used to determine
+// whether a cached entry is stil valid - if it was cached before the last write
+// to the store, it can't be trusted anymore.
 func (c *InMemoryCacheController) DetermineInvalidationTime(
 	ctx context.Context,
 	storeID string,
@@ -136,26 +151,37 @@ func (c *InMemoryCacheController) DetermineInvalidationTime(
 	defer span.End()
 	cacheTotalCounter.Inc()
 
+	// Changelog cache entry holds the last modified time for the store.
 	cacheKey := storage.GetChangelogCacheKey(storeID)
 	cacheResp := c.cache.Get(cacheKey)
+
 	c.logger.Debug("InMemoryCacheController DetermineInvalidationTime cache attempt",
 		zap.String("store_id", storeID),
 		zap.Bool("hit", cacheResp != nil),
 	)
-	if cacheResp != nil {
-		if entry, ok := cacheResp.(*storage.ChangelogCacheEntry); ok {
-			// the TTL grace period hasn't been breached
-			if entry.LastModified.Add(c.ttl).After(time.Now()) {
-				cacheHitCounter.Inc()
-				span.SetAttributes(attribute.Bool("cached", true))
-				return entry.LastModified
-			}
-		}
+
+	// entry is nil when cacheResp is nil (cache miss) or type assertion fails
+	// (shouldn't happen since we used the unique changelog cache key prefix
+	// when getting from cache).
+	entry, _ := cacheResp.(*storage.ChangelogCacheEntry)
+	if entry == nil {
+		c.InvalidateIfNeeded(ctx, storeID) // async
+
+		// Return zero time to allow caller to use cache while invalidation is in progress.
+		return time.Time{}
 	}
 
-	c.InvalidateIfNeeded(ctx, storeID)
+	// Cache hit
+	cacheHitCounter.Inc()
+	span.SetAttributes(attribute.Bool("cached", true))
 
-	return time.Time{}
+	// Ensure invalidation is triggered at most every c.ttl duration per store.
+	if time.Since(entry.LastChecked) > c.ttl {
+		c.InvalidateIfNeeded(ctx, storeID) // async
+	}
+
+	// Return time of last known change to store.
+	return entry.LastModified
 }
 
 // findChangesDescending is a wrapper on ReadChanges. If there are 0 changes to be returned, ReadChanges will actually return an error.
@@ -169,6 +195,9 @@ func (c *InMemoryCacheController) findChangesDescending(ctx context.Context, sto
 	return c.ds.ReadChanges(ctx, storeID, storage.ReadChangesFilter{}, opts)
 }
 
+// InvalidateIfNeeded checks to see if an invalidation is currently in progress for a store,
+// and if not it will spawn a goroutine to invalidate cached records conditionally
+// based on timestamp. It may invalidate all cache records, some, or none.
 func (c *InMemoryCacheController) InvalidateIfNeeded(ctx context.Context, storeID string) {
 	span := trace.SpanFromContext(ctx)
 	_, present := c.inflightInvalidations.LoadOrStore(storeID, struct{}{})
@@ -251,10 +280,13 @@ func (c *InMemoryCacheController) findChangesAndInvalidateIfNecessary(parentCtx 
 	lastChangelog := changes[0]
 	entry := &storage.ChangelogCacheEntry{
 		LastModified: lastChangelog.GetTimestamp().AsTime(),
+		LastChecked:  time.Now(),
 	}
 
-	defer c.cache.Set(cacheKey, entry, utils.JitterDuration(c.ttl, time.Minute)) // add buffer between checks
-
+	// The changelog cache entry is only used to compare against a cached Check
+	// response. Therefore, we only need this entry for up to the TTL of the
+	// cached Check response (queryCacheTTL).
+	c.cache.Set(cacheKey, entry, c.queryCacheTTL)
 	invalidationType := "none"
 
 	if !lastChangelog.GetTimestamp().AsTime().After(lastInvalidation) {